pyproject.toml
[project]
name = "devops-final-backend"
version = "0.2.0"
description = "FastAPI Backend for the DevOps Final Project"
readme = "README.md"
authors = [
    { name = "Tudor Huza", email = "huzatudor@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.118.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.37.0",
    "langchain==0.3.27",
    "langchain-ollama>=0.3.10",
    "langchain-community>=0.3.30",
    "python-keycloak[openid]>=5.8.1",
]

[project.scripts]
devops-final-backend = "devops_final_backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
fix = true
lint.extend-select = ["I"]
